+++ lib/archive.js
@@ -3,7 +3,7 @@ var _ = require("underscore");
 var fs = require("fs.extra");
 var treeHash = require("treehash");
 var blockStream = require("block-stream");
-var timber = require("../timber");
+var timber = require("./timber");
 
 module.exports = {
   checkGlacierVaults: function (dirs, vault, callback) {
