+++ lib/handle.js
@@ -8,7 +8,7 @@ var _ = require("underscore");
 var url = require("url");
 var archiver = require("archiver");
 var async = require("async");
-var timber = require("../timber");
+var timber = require("./timber");
 var domain = require("domain");
 
 http.globalAgent.maxSockets = 5;
