+++ lib/parse.js
@@ -10,7 +10,7 @@ var querystring = require("querystring");
 var fs = require("fs");
 var path = require("path");
 var async = require("async");
-var handle = require("../handle");
+var handle = require("./handle");
 var domain = require("domain");
 
 module.exports = {
