+++ lib/timber.js
