+++ lib/utility.js
